drivers: led: pca9633: disable allcall
The PCA9633 i2c LED controller offers an All Call address in its nominal operation, allowing simultaneous communication with all instances present on the same i2c bus. The default address is 0x70. While this functionality is convenient, it is possible that the board uses another i2c component that also uses this address (for example, the shtcx). In such cases, the address conflict prevents the proper functioning of the system.

The idea is to add a "disable-allcall" property to the device tree. If this option is present, the initialization of the PCA9633 forces the bit 0 (ALLCALL) to be set to false, thereby disabling this function. It is necessary to add this property to all PCA9633 devices on the bus to free up the address 0x70.

Signed-off-by: Steve Jacot-Guillarmod <steve.jacot-guillarmod@swissdotnet.ch>